# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Private Practice Management & Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/bar-registration:
    put:
      operationId: setBarRegistration
      tags: [users]
      summary: Declare bar enrolment (self-declared, not verified against a bar council)
      description: >
        Required before an admin, senior_advocate or junior_advocate reaches
        the dashboard — see `SessionClaims.profileComplete`. Callable at any
        time afterward to correct or update what was declared; there is no
        separate admin approval step, because this is self-declaration, not a
        credential check.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarRegistrationInput"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarRegistration"
        "400":
          description: Invalid
        "401":
          description: Unauthorized

  # ── Session & workspaces ───────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, senior_advocate, junior_advocate, clerk_intern, clerk, client]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  # ── Cases ──────────────────────────────────────────────────────────────────
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [time]
      summary: Time logged, optionally for one matter
      parameters:
        - in: query
          name: caseId
          schema: { type: integer }
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeEntry" }
        "403": { description: Forbidden }
    post:
      operationId: createTimeEntry
      tags: [time]
      summary: Log time against a matter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimeEntryInput" }
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeEntry" }
        "400": { description: Invalid }
        "403": { description: Forbidden }
  /time-entries/{id}:
    delete:
      operationId: deleteTimeEntry
      tags: [time]
      summary: Delete an entry you logged
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
        "403": { description: Forbidden }
        "404": { description: Not found }
  /time-entries/timer:
    get:
      operationId: getRunningTimer
      tags: [time]
      summary: The caller's running timer, if any
      responses:
        "200":
          description: Running timer or null
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunningTimer" }
        "403": { description: Forbidden }
    post:
      operationId: startTimer
      tags: [time]
      summary: Start a timer on a matter
      description: >
        One timer per person. Starting a second stops the first and banks its
        minutes, because the alternative is silently double-counting.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimerStartInput" }
      responses:
        "201":
          description: Started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunningTimer" }
        "400": { description: Invalid }
        "403": { description: Forbidden }
    delete:
      operationId: stopTimer
      tags: [time]
      summary: Stop the running timer and bank the minutes
      responses:
        "200":
          description: The finished entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeEntry" }
        "404": { description: No timer running }
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, role, roleSelected, displayName, email, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        role:
          type: string
          enum: [admin, senior_advocate, junior_advocate, clerk_intern, clerk, client]
        roleSelected:
          type: boolean
          description: True once the user has completed self-service role selection at sign-up.
        displayName:
          type: string
        email:
          type: string
        authProvider:
          type: ["string", "null"]
        barCouncilState:
          type: ["string", "null"]
        barEnrolmentNo:
          type: ["string", "null"]
        aorNo:
          type: ["string", "null"]
          description: Advocate-on-Record, Supreme Court of India.
        aorHighCourtNo:
          type: ["string", "null"]
          description: Advocate-on-Record at a High Court, where that court keeps a roll.
        copNo:
          type: ["string", "null"]
          description: Certificate of Practice number.
        allIndiaBarNo:
          type: ["string", "null"]
          description: All India Bar Examination certificate number.
        allIndiaBarDueAt:
          type: ["string", "null"]
          format: date-time
          description: >
            When the All India Bar number stops being optional. Stamped once, on
            the first declaration, and never moved — a deadline that resets each
            time the form is saved is not a deadline.
        allIndiaBarDaysLeft:
          type: ["integer", "null"]
          description: >
            Days until it is required. Null once supplied or when no deadline is
            set; negative once overdue.
        barDeclaredAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
          description: >
            What to call this person. Stored as given and shown to their own
            chamber on every record they touch, so it is bounded rather than
            free: 120 characters holds an Indian name with honorifics and
            nothing that would break a ledger row's layout. Empty is refused —
            the way to have no name is to never set one, not to clear it.

    # ── Session & workspaces ──────────────────────────────────────────────
...
    BarRegistrationInput:
      type: object
      required: [barCouncilState, barEnrolmentNo]
      properties:
        barCouncilState:
          type: string
          minLength: 1
          description: >
            Self-declared, loosely validated — enrolment formats vary by
            state bar and are not standardised. What is typed is stored.
        barEnrolmentNo:
          type: string
          minLength: 1
        aorNo:
          type: string
          description: Supreme Court Advocate-on-Record number. Optional — most advocates never hold one.
        aorHighCourtNo:
          type: string
          description: >
            Advocate-on-Record at a High Court, where that court keeps such a
            roll. A different roll from the Supreme Court one, so a separate
            field — an advocate may hold either, both or neither.
        copNo:
          type: string
          description: Certificate of Practice number, issued by the state bar council.
        allIndiaBarNo:
          type: string
          description: >
            All India Bar Examination certificate number. Optional at
            declaration and required within six months of it — an advocate
            enrolled before the examination existed may hold none, and a newly
            enrolled one has a window in which to sit it.
    BarRegistration:
      type: object
      required: [barCouncilState, barEnrolmentNo, barDeclaredAt]
      properties:
        barCouncilState:
          type: string
        barEnrolmentNo:
          type: string
        aorNo:
          type: ["string", "null"]
        aorHighCourtNo:
          type: ["string", "null"]
        copNo:
          type: ["string", "null"]
        allIndiaBarNo:
          type: ["string", "null"]
        allIndiaBarDueAt:
          type: ["string", "null"]
          format: date-time
          description: When the All India Bar number stops being optional.
        allIndiaBarDaysLeft:
          type: ["integer", "null"]
          description: >
            Days until it is required. Null once supplied or when no deadline is
            set; negative once overdue. Drives the warning before the gate bites.
        barDeclaredAt:
          type: string
          format: date-time
...
    TimeEntry:
      type: object
      required: [id, caseId, caseTitle, userId, userName, workDate, minutes, billable, createdAt]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        caseTitle: { type: string }
        userId: { type: integer }
        userName: { type: string }
        workDate:
          {
            type: string,
            pattern: "^\\d{4}-\\d{2}-\\d{2}$",
            description: "Calendar date, YYYY-MM-DD.",
          }
        minutes: { type: integer, description: Exact minutes. Never a fractional hour. }
        description: { type: ["string", "null"] }
        billable: { type: boolean }
        startedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    TimeEntryInput:
      type: object
      required: [caseId, workDate, minutes]
      properties:
        caseId: { type: integer }
        workDate:
          {
            type: string,
            pattern: "^\\d{4}-\\d{2}-\\d{2}$",
            description: "Calendar date, YYYY-MM-DD.",
          }
        minutes: { type: integer, minimum: 1, maximum: 1440 }
        description: { type: string, maxLength: 2000 }
        billable: { type: boolean, default: true }
    TimerStartInput:
      type: object
      required: [caseId]
      properties:
        caseId: { type: integer }
    RunningTimer:
      type: ["object", "null"]
      required: [id, caseId, caseTitle, startedAt, elapsedMinutes]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        caseTitle: { type: string }
        startedAt: { type: string, format: date-time }
        elapsedMinutes: { type: integer }
...